feat: fixing logo screen performance issue
